requirements.txt
langchain-google-genai
python-dotenv
python-telegram-bot
requests
